requirements.txt
azure-functions

numpy
pandas
